+++ mapsforge-map-android/src/main/java/org/mapsforge/map/android/view/MapView.java
@@ -19,6 +19,7 @@ package org.mapsforge.map.android.view;
 import org.mapsforge.core.graphics.GraphicFactory;
 import org.mapsforge.core.model.BoundingBox;
 import org.mapsforge.core.model.Dimension;
+import org.mapsforge.core.model.LatLong;
 import org.mapsforge.map.android.graphics.AndroidGraphicFactory;
 import org.mapsforge.map.android.input.MapZoomControls;
 import org.mapsforge.map.android.input.ScaleListener;
@@ -251,4 +252,19 @@ public class MapView extends ViewGroup implements org.mapsforge.map.view.MapView
 		}
 		this.mapScaleBar = mapScaleBar;
 	}
+
+
+	// NW convenience methods to simplify API
+	public void setCenter(LatLong latLong) {
+		this.model.mapViewPosition.setCenter(latLong);
+	}
+
+	public void setZoomLevel(byte zoomLevel) {
+		this.model.mapViewPosition.setZoomLevel(zoomLevel);
+	}	
+
+	public void addLayer(Layer layer) {
+		this.layerManager.getLayers().add(layer);
+	}		
+	// NW convenience methods end
 }