+++ mapsforge-map/src/main/java/org/mapsforge/map/view/MapView.java
@@ -17,9 +17,11 @@ package org.mapsforge.map.view;
 
 import org.mapsforge.core.model.BoundingBox;
 import org.mapsforge.core.model.Dimension;
+import org.mapsforge.core.model.LatLong;
 import org.mapsforge.map.layer.LayerManager;
 import org.mapsforge.map.model.Model;
 import org.mapsforge.map.scalebar.MapScaleBar;
+import org.mapsforge.map.layer.Layer;
 
 public interface MapView {
 
@@ -61,4 +63,9 @@ public interface MapView {
 	void repaint();
 
 	void setMapScaleBar(MapScaleBar mapScaleBar);
+
+	// NW convenience methods
+	public void setCenter(LatLong latLong);
+	public void setZoomLevel(byte zoomLevel);
+	public void addLayer(Layer layer);
 }